registrationApi/requirements.txt
flask~=1.1
python-dotenv==0.15.0
mysql-connector-python~=8.0
redis~=3.5
requests~=2.25
gunicorn~=20.0
sphinx~=3.5